requirements.txt
# Hermes - Internet Speed Test Automation and Visualization
# Frozen with: pip freeze (Python 3.13, venv)
# To update: pip install -U <package> then re-freeze.

# --- Core ---
speedtest-cli==2.1.3
APScheduler==3.11.2
python-dotenv==1.2.2
tzlocal==5.3.1

# --- Exporters ---
prometheus_client==0.24.1
requests==2.33.0
types-requests==2.33.0.20260402

# --- REST API ---
fastapi==0.115.14
uvicorn[standard]==0.34.3
httpx==0.28.1
aiofiles==24.1.0

# --- Testing / dev ---
watchdog==6.0.0

# --- Development / testing (excluded from production Docker build) ---
pytest==9.0.2
iniconfig==2.3.0
pluggy==1.6.0
Pygments==2.19.2
ruff==0.15.8hypothesis==6.152.1